+++ phpunit/class-wp-rest-themes-controller-test.php
@@ -179,9 +179,10 @@ class WP_REST_Themes_Controller_Test extends WP_Test_REST_Controller_Testcase {
 		$this->assertEquals( 1, count( $properties ) );
 		$this->assertArrayHasKey( 'theme_supports', $properties );
 
-		$this->assertEquals( 2, count( $properties['theme_supports']['properties'] ) );
+		$this->assertEquals( 3, count( $properties['theme_supports']['properties'] ) );
 		$this->assertArrayHasKey( 'formats', $properties['theme_supports']['properties'] );
 		$this->assertArrayHasKey( 'post-thumbnails', $properties['theme_supports']['properties'] );
+		$this->assertArrayHasKey( 'responsive-embeds', $properties['theme_supports']['properties'] );
 	}
 
 	/**
@@ -206,6 +207,31 @@ class WP_REST_Themes_Controller_Test extends WP_Test_REST_Controller_Testcase {
 		$this->assertTrue( isset( $result[0]['theme_supports'] ) );
 		$this->assertTrue( isset( $result[0]['theme_supports']['formats'] ) );
 		$this->assertSame( array( 'standard', 'aside', 'video' ), $result[0]['theme_supports']['formats'] );
+	}
+
+	/**
+	 * Test when a theme does not support responsive embeds.
+	 */
+	public function test_theme_supports_responsive_embeds_false() {
+		remove_theme_support( 'responsive-embeds' );
+		$response = self::perform_active_theme_request();
+
+		$result = $response->get_data();
+		$this->assertTrue( isset( $result[0]['theme_supports'] ) );
+		$this->assertTrue( isset( $result[0]['theme_supports']['responsive-embeds'] ) );
+		$this->assertFalse( $result[0]['theme_supports']['responsive-embeds'] );
+	}
+
+	/**
+	 * Test when a theme supports responsive embeds.
+	 */
+	public function test_theme_supports_responsive_embeds_true() {
+		remove_theme_support( 'responsive-embeds' );
+		add_theme_support( 'responsive-embeds' );
+		$response = self::perform_active_theme_request();
+		$result   = $response->get_data();
+		$this->assertTrue( isset( $result[0]['theme_supports'] ) );
+		$this->assertTrue( $result[0]['theme_supports']['responsive-embeds'] );
 	}
 
 	/**