+++ packages/block-library/src/list/index.js
@@ -85,7 +85,7 @@ export const settings = {
 				blocks: [ 'core/quote' ],
 				transform: ( { value } ) => {
 					return createBlock( 'core/list', {
-						values: value,
+						values: toHTMLString( create( { html: value, multilineTag: 'p' } ), 'li' ),
 					} );
 				},
 			},
@@ -111,7 +111,7 @@ export const settings = {
 				regExp: /^[*-]\s/,
 				transform: ( { content } ) => {
 					return createBlock( 'core/list', {
-						values: content,
+						values: `<li>${ content }</li>`,
 					} );
 				},
 			},
@@ -121,7 +121,7 @@ export const settings = {
 				transform: ( { content } ) => {
 					return createBlock( 'core/list', {
 						ordered: true,
-						values: content,
+						values: `<li>${ content }</li>`,
 					} );
 				},
 			},
@@ -143,7 +143,7 @@ export const settings = {
 				blocks: [ 'core/quote' ],
 				transform: ( { values } ) => {
 					return createBlock( 'core/quote', {
-						value: values,
+						value: toHTMLString( create( { html: values, multilineTag: 'li' } ), 'p' ),
 					} );
 				},
 			},