+++ packages/edit-post/src/store/selectors.js
@@ -1,12 +1,13 @@
 /**
+ * External dependencies
+ */
+import createSelector from 'rememo';
+import { get, includes, some, flatten, values } from 'lodash';
+
+/**
  * WordPress dependencies
  */
 import deprecated from '@wordpress/deprecated';
-
-/**
- * External dependencies
- */
-import { get, includes } from 'lodash';
 
 /**
  * Returns the current editing mode.
@@ -104,16 +105,49 @@ export function isPublishSidebarOpened( state ) {
 }
 
 /**
- * Returns true if the editor sidebar panel is open, or false otherwise.
+ * Returns true if the given panel is enabled, or false otherwise. Panels are
+ * enabled by default.
  *
- * @param  {Object}  state Global application state.
- * @param  {string}  panel Sidebar panel name.
+ * @param {Object} state     Global application state.
+ * @param {string} panelName A string that identifies the panel.
  *
- * @return {boolean} Whether the sidebar panel is open.
+ * @return {boolean} Whether or not the panel is enabled.
+ */
+export function isEditorPanelEnabled( state, panelName ) {
+	const panels = getPreference( state, 'panels' );
+	return get( panels, [ panelName, 'enabled' ], true );
+}
+
+/**
+ * Returns true if the given panel is enabled, or false otherwise. Panels are
+ * enabled by default.
+ *
+ * @param {Object} state Global application state.
+ * @param {string} panel A string that identifies the panel.
+ *
+ * @return {boolean} Whether or not the panel is enabled.
  */
 export function isEditorSidebarPanelOpened( state, panel ) {
+	deprecated( 'isEditorSidebarPanelOpened', {
+		alternative: 'isEditorPanelEnabled',
+		plugin: 'Gutenberg',
+		version: '4.3',
+	} );
+	return isEditorPanelEnabled( state, panel );
+}
+
+/**
+ * Returns true if the given panel is open, or false otherwise. Panels are
+ * closed by default.
+ *
+ * @param  {Object}  state     Global application state.
+ * @param  {string}  panelName A string that identifies the panel.
+ *
+ * @return {boolean} Whether or not the panel is open.
+ */
+export function isEditorPanelOpened( state, panelName ) {
 	const panels = getPreference( state, 'panels' );
-	return panels ? !! panels[ panel ] : false;
+	return panels[ panelName ] === true || get( panels, [ panelName, 'opened' ], false );
 }
 
 /**
@@ -189,8 +223,31 @@ export function getMetaBoxes( state ) {
  *
  * @return {string[]} Active meta box locations.
  */
-export function getActiveMetaBoxLocations( state ) {
-	return state.activeMetaBoxLocations;
+export const getActiveMetaBoxLocations = createSelector(
+	( state ) => {
+		return Object.keys( state.metaBoxes.locations )
+			.filter( ( location ) => isMetaBoxLocationActive( state, location ) );
+	},
+	( state ) => [
+		state.metaBoxes.locations,
+	]
+);
+
+/**
+ * Returns true if a metabox location is active and visible
+ *
+ * @param {Object} state    Post editor state.
+ * @param {string} location Meta box location to test.
+ *
+ * @return {boolean} Whether the meta box location is active and visible.
+ */
+export function isMetaBoxLocationVisible( state, location ) {
+	return (
+		isMetaBoxLocationActive( state, location ) &&
+		some( getMetaBoxesPerLocation( state, location ), ( { id } ) => {
+			return isEditorPanelEnabled( state, `meta-box-${ id }` );
+		} )
+	);
 }
 
 /**
@@ -203,8 +260,37 @@ export function getActiveMetaBoxLocations( state ) {
  * @return {boolean} Whether the meta box location is active.
  */
 export function isMetaBoxLocationActive( state, location ) {
-	return getActiveMetaBoxLocations( state ).includes( location );
+	const metaBoxes = getMetaBoxesPerLocation( state, location );
+	return !! metaBoxes && metaBoxes.length !== 0;
 }
+
+/**
+ * Returns the list of all the available meta boxes for a given location.
+ *
+ * @param {Object} state    Global application state.
+ * @param {string} location Meta box location to test.
+ *
+ * @return {?Array} List of meta boxes.
+ */
+export function getMetaBoxesPerLocation( state, location ) {
+	return state.metaBoxes.locations[ location ];
+}
+
+/**
+ * Returns the list of all the available meta boxes.
+ *
+ * @param {Object} state Global application state.
+ *
+ * @return {Array} List of meta boxes.
+ */
+export const getAllMetaBoxes = createSelector(
+	( state ) => {
+		return flatten( values( state.metaBoxes.locations ) );
+	},
+	( state ) => [
+		state.metaBoxes.locations,
+	]
+);
 
 /**
  * Returns the state of legacy meta boxes.
@@ -243,5 +329,5 @@ export function hasMetaBoxes( state ) {
  * @return {boolean} Whether the metaboxes are being saved.
  */
 export function isSavingMetaBoxes( state ) {
-	return state.isSavingMetaBoxes;
+	return state.metaBoxes.isSaving;
 }