+++ packages/editor/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - `wp.editor.PanelColor` has been deprecated in favor of `wp.editor.PanelColorSettings`.
 
+### New Features
+
+- Added `onClose` prop to `URLPopover` component.
+
 ## 4.0.0 (2018-09-30)
 
 ### Breaking Changes